Extract ReaderViewModel event handling into a seperate EventChannel object
