+++ r2-testapp/src/main/java/org/readium/r2/testapp/audiobook/AudiobookActivity.kt
@@ -33,7 +33,7 @@ class AudiobookActivity : R2AudiobookActivity() {
         super.onCreate(savedInstanceState)
 
         ViewModelProvider(this).get(ReaderViewModel::class.java)
-            .subscribeEvents(this) {
+            .channel.receive(this) {
                 when(it) {
                     is ReaderViewModel.Event.OpenOutlineRequested -> showOutlineFragment()
                     is ReaderViewModel.Event.OpenDrmManagementRequested -> showDrmManagementFragment()
